Fix ExifTool test failures and Perl5 core test regressions
- Always try AST splitting before interpreter fallback on MethodTooLargeException
- Disable codegen-time block refactoring that broke control flow with goto labels
- Fix BitwiseOperators int wrapping for 32-bit operations
- Fix regex state save/restore for statement modifier while/for
- Fix Long.MIN_VALUE negation overflow in shift operations
- Add native JNA getpwnam/getpwuid/getpwent support
- Return child PID from piped open, support waitpid for Java processes
- Fix ctime format to match Perl 5 output

Test results (matching master):
- re/pat.t: 1055, io/open.t: 183-184, op/bop.t: 488
- ExifTool: 590/600 (98%)

+++ src/main/java/org/perlonjava/runtime/operators/MathOperators.java
@@ -433,17 +433,14 @@ public class MathOperators {
      * @return A new RuntimeScalar representing the integer division result.
      */
     public static RuntimeScalar integerDivide(RuntimeScalar arg1, RuntimeScalar arg2) {
-        // Convert to integers and perform integer division
-        int dividend = arg1.getInt();
-        int divisor = arg2.getInt();
+        long dividend = arg1.getLong();
+        long divisor = arg2.getLong();
 
-        // Handle division by zero
         if (divisor == 0) {
             throw new PerlCompilerException("Illegal division by zero");
         }
 
-        // Perform integer division
-        int result = dividend / divisor;
+        long result = dividend / divisor;
         return new RuntimeScalar(result);
     }
 
@@ -456,18 +453,14 @@ public class MathOperators {
      * @return A new RuntimeScalar representing the integer modulus.
      */
     public static RuntimeScalar integerModulus(RuntimeScalar arg1, RuntimeScalar arg2) {
-        // Convert to integers and perform integer modulus
-        int dividend = arg1.getInt();
-        int divisor = arg2.getInt();
+        long dividend = arg1.getLong();
+        long divisor = arg2.getLong();
 
-        // Handle division by zero
         if (divisor == 0) {
             throw new PerlCompilerException("Illegal modulus zero");
         }
 
-        // In Perl with "use integer", modulus follows C99 truncated division rules
-        // The result has the sign of the dividend, not the divisor
-        int result = dividend % divisor;
+        long result = dividend % divisor;
         return new RuntimeScalar(result);
     }
